requirements.txt
aiofiles==24.1.0 ; python_version >= "3.10" and python_version < "3.15"
aiogram==3.15.0 ; python_version >= "3.10" and python_version < "3.15"
aiohappyeyeballs==2.4.4 ; python_version >= "3.10" and python_version < "3.15"
aiohttp==3.10.11 ; python_version >= "3.10" and python_version < "3.15"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "3.15"
aiosqlite==0.20.0 ; python_version >= "3.10" and python_version < "3.15"
annotated-types==0.7.0 ; python_version >= "3.10" and python_version < "3.15"
async-timeout==4.0.3 ; python_version == "3.10"
attrs==23.2.0 ; python_version >= "3.10" and python_version < "3.15"
certifi==2024.8.30 ; python_version >= "3.10" and python_version < "3.15"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.15" and sys_platform == "win32"
colorlog==6.8.2 ; python_version >= "3.10" and python_version < "3.15"
frozenlist==1.5.0 ; python_version >= "3.10" and python_version < "3.15"
google-colab ; python_version >= "3.10" and sys_platform == "linux"
idna==3.10 ; python_version >= "3.10" and python_version < "3.15"
magic-filter==1.0.12 ; python_version >= "3.10" and python_version < "3.15"
multidict==6.1.0 ; python_version >= "3.10" and python_version < "3.15"
propcache==0.2.0 ; python_version >= "3.10" and python_version < "3.15"
pydantic-core==2.23.2 ; python_version >= "3.10" and python_version < "3.15"
pydantic==2.9.0 ; python_version >= "3.10" and python_version < "3.15"
python-dotenv==1.0.1 ; python_version >= "3.10" and python_version < "3.15"
typing-extensions==4.14.0 ; python_version >= "3.10" and python_version < "3.15"
typing-inspection==0.4.0 ; python_version >= "3.10" and python_version < "3.15"
yarl==1.17.0 ; python_version >= "3.10" and python_version < "3.15"